chore: rename history view managers
