+++ src/historyView/index.html
@@ -20,9 +20,9 @@
         "imports": {
           "./modules/domHandlers.js": "${domHandlersUri}",
           "./modules/uiManagers/Events.js": "${eventsUri}",
-          "./modules/uiManagers/historyRenderer.js": "${historyRendererUri}",
+          "./modules/uiManagers/HistoryRenderer.js": "${historyRendererUri}",
           "./modules/historyViewState.js": "${historyViewStateUri}",
-          "./modules/uiManagers/searchManager.js": "${searchManagerUri}",
+          "./modules/uiManagers/SearchManager.js": "${searchManagerUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
           "@common/webviewState.js": "${webviewStateUri}",
           "@common/webview/commands.js": "${commandsUri}",
