+++ src/historyView/modules/domHandlers.js
@@ -1,6 +1,6 @@
 // Local imports
-import { HistoryRenderer } from './uiManagers/historyRenderer.js';
-import { SearchManager } from './uiManagers/searchManager.js';
+import { HistoryRenderer } from './uiManagers/HistoryRenderer.js';
+import { SearchManager } from './uiManagers/SearchManager.js';
 import { HistoryEvents } from './uiManagers/Events.js';
 import { historyViewState } from './historyViewState.js';
