+++ src/historyView/modules/uiManagers/HistoryRenderer.js
